checks fro equality before overwriting existing values
git-svn-id: http://192.168.1.10/repos/toto@1542 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/DatabaseServer/src/com/azquo/memorydb/service/ValueService.java
@@ -118,6 +118,9 @@ public final class ValueService {
 
 
     public boolean overWriteExistingValue(final AzquoMemoryDBConnection azquoMemoryDBConnection, final Value existingValue, final String newValueString) throws Exception {
+        if (newValueString.equals(existingValue.getText())){
+            return true;
+        }
         Value newValue = new Value(azquoMemoryDBConnection.getAzquoMemoryDB(), azquoMemoryDBConnection.getProvenance(), newValueString);
         newValue.setNamesWillBePersisted(new HashSet<>(existingValue.getNames())); // a bit crappy but I'm trying to make it a list internally but interfaced by sets
         existingValue.delete();
